+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
   **CHANGELOG**
   - Added font size control
+  - Added keyboard friendly form into resource page for making reservations
 
 # 0.10.0
   **MINOR CHANGES**